Use Stdlib.Seq module rather than bespoke Stream module (#3)
* Remove OCaml version from build command

This is what recent versions of dune generate and it matches upstream changes in the opam repository as well.

* Add qcheck tests for Zar.take

Add tests for assumptions of Zar.take. Also, renamed the test file to test_zar to fix the dune build errors.

* Separate basic tests

Convert the sampler build, run, and print tests to dune expect style tests.

* Initial move to Seq

Step 1 of converting stream to seq module--it keeps the same interface but implements it with the seq module.  The expect file must be updated as the behavior is slightly different.

* Don't include the Stream module

* Don't open the Stream module

* Replace Stream with Seq

* Drop the stream module

* Include Seq_ext functions in Zar module

This change is to match the original behavior, where the stream functions were made available in the Zar module.  Not all of them are needed now that most of the "stream" behavior is covered by the Seq module itself.

* Don't reference the internal modules

* Use the new function name